Add unique constraint to ExternalField in MappingRules
This commit introduces a unique constraint on the `ExternalField` column in the `MappingRules` table by adding an index that enforces uniqueness. Changes are made in the `MappingRuleConfiguration` class and the `MappingRuleContextModelSnapshot`. A new migration class `AddUniqueConstraintToExternalFieldCol` is created to manage the database migration, including methods to create and drop the unique index. The designer file for the migration is also updated accordingly.